Can now parse mapfiles written entirely on a single line, as well as with newlines.
The new grammar file is even a little simpler now. However, it is x3 times slower than the newline-sensitive grammar. The reason is that the newlines provided some ambiguity resolution. Without it, the parser has to resolve a lot more branches. If the performance degradation is too significant for you, I can suggest a few ways to improve it. But no guarantees.